refactor(1682): migrate prod call sites off the MCP shims to reyn.mcp.*
Point the 9 production importers at the new submodule paths (the shims remain for
tests + any external callers; removing them is the optional follow-up):

  server (5): web/routers/a2a.py, web/routers/mcp.py, cli/commands/dogfood.py,
              cli/commands/chat.py, cli/commands/mcp.py  → reyn.mcp.server
  client (4): chat/session.py, op_runtime/mcp.py, cli/commands/mcp.py (×2)
              → reyn.mcp.client

Touched only the mcp-import lines (session.py:4901, cli/commands/chat.py:49) —
disjoint from tui-coder's parallel #2 regions; git should auto-merge.

Refs #1682
